+++ app/src/main/java/com/example/goodstuff/MyListData.java
@@ -8,15 +8,15 @@ import java.util.Date;
 
 public class MyListData {
     private final String day, date;
-    private long hours;
+    private double hours;
 
-    public MyListData(String day, String date, long hours) {
+    public MyListData(String day, String date, double hours) {
         this.day = day;
         this.date = date;
         this.hours = hours;
     }
 
-    public static MyListData construct(Date date, long hours) {
+    public static MyListData construct(Date date, double hours) {
         @SuppressLint("SimpleDateFormat") String dy = (new SimpleDateFormat("EE")).format(date);
         @SuppressLint("SimpleDateFormat") String dt = (new SimpleDateFormat("dd-MM-yyyy")).format(date);
         return new MyListData(dy, dt, hours);
@@ -30,11 +30,11 @@ public class MyListData {
         return this.date;
     }
 
-    public long getHours() {
+    public double getHours() {
         return hours;
     }
 
-    public void setHours(long hours) {
+    public void setHours(Double hours) {
         this.hours = hours;
     }
 }
